final-pjt/back-server/requirements.txt
appnope==0.1.3
asgiref==3.6.0
asttokens==2.2.1
backcall==0.2.0
beautifulsoup4==4.9.3
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.4
cryptography==40.0.2
decorator==5.0.7
defusedxml==0.7.1
dj-rest-auth==4.0.0
Django==3.2.13
django-allauth==0.52.0
django-bootstrap-v5==1.0.2
django-cors-headers==3.14.0
django-extensions==3.1.3
django-filter==23.2
django-seed==0.2.2
djangorestframework==3.13.1
executing==1.2.0
Faker==8.1.1
idna==3.4
importlib-metadata==2.1.1
ipython==7.22.0
ipython-genutils==0.2.0
jedi==0.18.0
joblib==1.2.0
matplotlib-inline==0.1.3
numpy==1.24.3
oauthlib==3.2.2
pandas==2.0.1
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
prompt-toolkit==3.0.18
psycopg2==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.8.1
PyJWT==2.7.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.1
requests==2.28.2
requests-oauthlib==1.3.1
scikit-learn==1.2.2
scipy==1.10.1
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
stack-data==0.6.2
text-unidecode==1.3
threadpoolctl==3.1.0
toposort==1.7
traitlets==5.0.5
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
wcwidth==0.2.5
zipp==3.4.1
